Move the form contents to separate component
